capacity-bench: share the traversal sweep with the engine and scale the profiles
Two problems with the traversal number, both real.

The tool was measuring the wrong thing. GET /benchmark traverses in parallel
across vertex ranges; the tool walked one edge group serially, so it reported a
slower number while the page claimed both did the same work. And the graph was
too small to measure at all: 100,000 edges completes in under a millisecond, so
successive passes on one machine reported 186M through 457M edges/s. That was
stopwatch resolution and thread-pool ramp, not throughput.

The sweep now lives in the engine as OutEdgeSweep and both callers use it, so a
number measured through REST and a number measured by the tool describe one code
path. The engine is the only place it can live: an out-of-assembly caller has to
reach adjacency through VertexModel.GetOutgoingEdgeIds(), which allocates a key
list per vertex. On a ten-million-vertex graph that was ten million allocations
per pass, inside the measurement. Inside the engine the allocation-free struct
enumerator over the groups is reachable, and the inner loop hoists the array,
count and offset so it carries no field loads.

That also speeds up GET /benchmark, which loses the same per-vertex allocation.

Measured on a 16-logical-core Intel workstation, best pass of five:

  5M edges    393M edges/s
  20M edges   371M edges/s
  100M edges  312M edges/s

The previous serial implementation reported 223M on a graph small enough to be
noise. The drop across the sizes is memory latency: following an edge is a chain
of dependent loads (adjacency slot, edge object, neighbour), so once the working
set outgrows cache each one is a round trip. Traversal is therefore reported at
several sizes rather than one, because a single figure either flatters the engine
on a small graph or understates it on a large one.

Profiles are scaled roughly tenfold. The full profile's headline shape is now
10,000,000 vertices with 100,000,000 edges, which retains about 13 GB and wants
32 GB or more; memory, save-stall and write scenarios grew with it. Graph
construction is batched (one million vertices, two million edges per
transaction), because a single transaction holds a definition object per element
until it commits and one for a hundred million edges is hundreds of megabytes on
top of the graph it is building.

Write scenarios take a time cap as well as a count: serial single-element writes
are fsync-bound at a few hundred per second, so the full profile's count would
otherwise add minutes for a number that is stable in seconds. The rate is
computed over what actually committed and the report and page now show that count
per row, so a capped run is still a checkable rate.

Adds --only <family> for iterating on one measurement without rebuilding a
hundred-million-edge graph. A filtered report is deliberately not publishable:
the renderer requires every family and refuses one that is missing.

+++ fallen-8-bench/CapacityReport.cs
@@ -144,7 +144,7 @@ namespace NoSQL.GraphDB.Bench
         public Int32 Vertices { get; set; }
 
         [JsonPropertyName("edges")]
-        public Int32 Edges { get; set; }
+        public Int64 Edges { get; set; }
 
         [JsonPropertyName("averageDegree")]
         public Double AverageDegree { get; set; }
@@ -169,7 +169,7 @@ namespace NoSQL.GraphDB.Bench
         public Int32 Producers { get; set; }
 
         [JsonPropertyName("writes")]
-        public Int32 Writes { get; set; }
+        public Int64 Writes { get; set; }
 
         [JsonPropertyName("writesPerSecond")]
         public Double WritesPerSecond { get; set; }
@@ -182,13 +182,13 @@ namespace NoSQL.GraphDB.Bench
     public sealed class SaveStallMetric
     {
         [JsonPropertyName("elements")]
-        public Int32 Elements { get; set; }
+        public Int64 Elements { get; set; }
 
         [JsonPropertyName("vertices")]
         public Int32 Vertices { get; set; }
 
         [JsonPropertyName("edges")]
-        public Int32 Edges { get; set; }
+        public Int64 Edges { get; set; }
 
         [JsonPropertyName("writerHoldMs")]
         public Double WriterHoldMs { get; set; }
@@ -204,7 +204,7 @@ namespace NoSQL.GraphDB.Bench
         public Int32 Vertices { get; set; }
 
         [JsonPropertyName("edges")]
-        public Int32 Edges { get; set; }
+        public Int64 Edges { get; set; }
 
         [JsonPropertyName("iterations")]
         public Int32 Iterations { get; set; }
